Add basic serializer for User model
In this initial version of the serializer, we want to simply limit
the set of attributes included in the serialized JSON object (by
removing :password_digest) and include the result of a method
(:gravatar_url) as a virtual attribute.

We also switch to using the :json adapter for AMS by default which
will wrap all model attributes in a root element with the same name
as the model class.