Add a :links attribute to the serialized User
We include a new attribute for passing along the URL for accessing
the serialized User model directly as well as the link for all
Boards belonging to the User.

In order to make this work as shown in the text, we need to update
BoardsController and the routes file to include a nested resource
for accessing the Boards belonging to a given User.