Add related Boards to User serializer
Declares two new :has_many associations on the User serializer to
include the arrays of :active_boards and :archived_boards created
by a given User, respectively.

To do this, we also needed to define a new BoardPreviewSerializer
which provides an abbreviated set of attributes for a given Board.
In order to keep a fuller set of attributes supplied to the endpoints
for the Board resource, we'll also need a default BoardSerializer.

Finally, we'll control which related data is sent through to clients
of the User-related endpoints by way of options passed to the render
methods in UsersController.