Test the :links included by UserSerializer
Write a test that demonstrates that the links included in the
serialized representation of a User model are the correct ones -
one for the User itself and another for the Boards created by that
User.

Also, extract the logic for creating the serialized representation
of the user to a private helper method.