Require HTTP basic authentication for all endpoints
Protect all controller actions with mandatory HTTP basic authentication.
The username and password should be stored as environment variables,
and the requirement to authenticate should be conditional based on the
presence of the username variable.

After introducing authentication for all v3 controllers, every test
immediately fails, of course. Introduce a new AuthenticationHelpers
mixin for generating the basic authentication header for a given set
of credentials.